pyproject.toml
[project]
name = "ai44pt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "matplotlib>=3.4.0",
    "numpy>=1.21.0",
    "openai>=1.0.0",
    "openpyxl==3.1.5",
    "pandas>=1.3.0",
    "pymupdf>=1.18.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=0.19.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
demo = [
    "gradio>=6.0.1",
]
